refactor(067): separate presentation and ui
