+++ app/src/androidTest/java/org/cescfe/numpairs/ui/screen/GameScreenLargeOperandsTest.kt
@@ -1,4 +1,4 @@
-package org.cescfe.numpairs.feature.game
+package org.cescfe.numpairs.feature.game.ui
 
 import androidx.activity.ComponentActivity
 import androidx.compose.foundation.layout.Box
