+++ app/src/main/java/org/cescfe/numpairs/feature/game/presentation/GameUiState.kt
@@ -1,4 +1,4 @@
-package org.cescfe.numpairs.feature.game
+package org.cescfe.numpairs.feature.game.presentation
 
 import org.cescfe.numpairs.domain.puzzle.Expression
 import org.cescfe.numpairs.domain.puzzle.OperandSelectionChoice
