+++ app/src/main/java/org/cescfe/numpairs/feature/game/ui/GameScreenTestTags.kt
@@ -1,4 +1,4 @@
-package org.cescfe.numpairs.feature.game
+package org.cescfe.numpairs.feature.game.ui
 
 import org.cescfe.numpairs.domain.puzzle.Operator
