+++ app/src/test/java/org/cescfe/numpairs/ui/screen/GameViewModelOperatorAndTileStateTest.kt
@@ -1,4 +1,4 @@
-package org.cescfe.numpairs.feature.game
+package org.cescfe.numpairs.feature.game.presentation
 
 import org.cescfe.numpairs.domain.puzzle.Operator
 import org.junit.Assert.assertEquals
